requirements.txt
ujson
Crypto
pyDes==2.0.1
werkzeug
Flask==0.12.2
Flask-Cors==3.0.3
Flask-RESTful==0.3.6
Flask-JWT-Simple==0.0.3
Flask-SQLAlchemy==2.3.2
requests==2.18.4
passlib==1.7.1
scikit-learn==0.19.1
xlrd == 1.1.0
psycopg2 == 2.7.4
pycrypto==2.6.1
